pyproject.toml
[project]
name = "so-scrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.15",
    "beautifulsoup4>=4.13.4",
    "matplotlib>=3.10.5",
    "notebook>=7.4.4",
    "pandas>=2.3.1",
    "pymongo>=4.13.2",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "ruff>=0.12.7",
    "seaborn>=0.13.2",
    "selenium>=4.34.2",
]
[tool.pytest.ini_options]
pythonpath = ["src"]
markers = [
    "api: tests related to external API calls",
    "scrapper: tests related to scrappers"
]
